Fixes after merge of 'master' into '4_CBTest'

+++ InterfaceFactory/CommHandler4Modbus/src/main/java/communicator/common/impl/SGrDeviceBase.java
@@ -132,14 +132,14 @@ public abstract class SGrDeviceBase<
         throws GenDriverException {
 
         Optional<String> errorStr = Optional.ofNullable(dataPoint.getGenericAttributes())
-                .flatMap(attr -> Optional.ofNullable(attr.getMaxVal()).flatMap(val -> checkOutOfRange(values, val, Comparator.MAX)));
+                .flatMap(attr -> Optional.ofNullable(attr.getMaxVal()).flatMap(val -> checkOutOfRange(values, (BigDecimal) val, Comparator.MAX)));
 
         if (errorStr.isPresent()) {
             throw new GenDriverException(errorStr.get());
         }
 
         errorStr = Optional.ofNullable(dataPoint.getGenericAttributes())
-                .flatMap(attr -> Optional.ofNullable(attr.getMinVal()).flatMap(val -> checkOutOfRange(values, val, Comparator.MIN)));
+                .flatMap(attr -> Optional.ofNullable(attr.getMinVal()).flatMap(val -> checkOutOfRange(values, (BigDecimal) val, Comparator.MIN)));
 
         if (errorStr.isPresent()) {
             throw new GenDriverException(errorStr.get());
